requirements.txt
dbt-core>=1.7.0
dbt-duckdb>=1.7.0

